Add "patterns" parameter to set the number of iterations for PRNG tests.
Up to now all pseudo random tests (08BIT, 16BIT, 32BIT and BURST) filled
the complete RAM with 6 different pre-defined patterns. Depending on the
RAM size this can take a long time.
The new parameter "patterns" selects the number of patterns to use. A
value of 0 selects all available patterns. A value higher than 6 is
limited to 6.