Make UART and progress configurable on netX50 standalone.
Add 2 new values to the parameter field of the standalone version.
The first one defines the MMIO pins for the UART RX and TX signals.
The second one is a placeholder for the progress display.

At the moment the UART MMIO pins are only evaluated on the netX50.
The progress display can not be configured and is still fixed to
RDY/RUN blinking like in previous versions.